FastAPIProject/requirements.txt
fastapi
uvicorn
openai
python-dotenv
langchain
langchain-openai
langchain-community
faiss-cpu
numpy
pandas
pypdf2
python-multipart
pydantic
sentence-transformers
tiktoken
huggingface-hub
transformers
